requirements.txt
charset-normalizer==2.1.0
distlib==0.3.5
Django==4.0.4
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-phone-field==1.8.1
feedparser==6.0.10
filelock==3.7.1
idna==3.3
install==1.3.5
Pillow==9.1.1
platformdirs==2.5.2
pycparser==2.21
pynput==1.7.3
python-dateutil==2.7.3     
pytz==2022.1
regex==2022.6.2
requests==2.28.1
sgmllib3k==1.0.0
sqlparse==0.4.2
urllib3==1.26.11
pdfkit==1.0.0
wkhtmltopdf==0.2
pandas ==1.2.2
numpy==1.23.2
matplotlib==3.5.2
django-import-export==2.8.0